Change closing parentheses and brackets indentation behavior, add setting
The new indentation behavior will place the closing parenteses and brackets
onto the current indentation level of the section. The previous behavior
would place the closing symbol vertically aligned with the opening one. The
old behavior is available if the new opascal-vertical-paren-align variable
is set to true. See documentation for an example.